refactor(browserstack-service): inline app-automate detection in config.ts
Per review, minimise the public surface area of this PR. The
hasAppAutomateCapability / hasAnyAppAutomateCapability helpers in
util.ts are not consumed outside of config.ts, so collapse them into
private functions in config.ts and revert util.ts / util.test.ts to
match upstream/main exactly.

Detection logic and behaviour are unchanged; only the location moved.
The config.test.ts test surface is also trimmed to the essential
scenarios (options.app, web-only, appium:app, appium:options.app,
W3C firstMatch, multiremote object form).
