+++ packages/wdio-browserstack-service/tests/config.test.ts
@@ -8,22 +8,21 @@ const baseConfig = { framework: 'mocha', user: 'u', key: 'k' } as any
 
 describe('BrowserStackConfig appAutomate detection', () => {
     beforeEach(() => {
-        // reset singleton so each test gets a fresh instance
         ;(BrowserStackConfig as any)._instance = undefined
     })
 
-    it('treats web-only caps with no `app` option as automate', () => {
+    it('marks app_automate when options.app is set (existing behaviour)', () => {
+        const cfg = new BrowserStackConfig({ app: 'bs://abc' } as any, baseConfig)
+        expect(cfg.appAutomate).toBe(true)
+        expect(cfg.automate).toBe(false)
+    })
+
+    it('marks automate for web-only caps with no options.app', () => {
         const cfg = new BrowserStackConfig(baseOptions, baseConfig, [
             { browserName: 'chrome' },
         ] as any)
         expect(cfg.appAutomate).toBe(false)
         expect(cfg.automate).toBe(true)
-    })
-
-    it('marks app_automate when options.app is set, even without caps', () => {
-        const cfg = new BrowserStackConfig({ app: 'bs://abc' } as any, baseConfig)
-        expect(cfg.appAutomate).toBe(true)
-        expect(cfg.automate).toBe(false)
     })
 
     it('marks app_automate when capabilities carry appium:app', () => {
@@ -34,75 +33,25 @@ describe('BrowserStackConfig appAutomate detection', () => {
         expect(cfg.automate).toBe(false)
     })
 
-    it('marks app_automate when capabilities carry appium:bundleId', () => {
+    it('marks app_automate for the nested appium:options.app form', () => {
         const cfg = new BrowserStackConfig(baseOptions, baseConfig, [
-            { platformName: 'iOS', 'appium:bundleId': 'com.example.app' },
-        ] as any)
-        expect(cfg.appAutomate).toBe(true)
-        expect(cfg.automate).toBe(false)
-    })
-
-    it('marks app_automate for android caps with appPackage', () => {
-        const cfg = new BrowserStackConfig(baseOptions, baseConfig, [
-            { platformName: 'Android', 'appium:appPackage': 'com.example' },
+            { 'appium:options': { app: 'bs://abc' } },
         ] as any)
         expect(cfg.appAutomate).toBe(true)
     })
 
-    it('marks app_automate for parallel multiremote with one app cap', () => {
+    it('marks app_automate when only a W3C firstMatch entry carries the app cap', () => {
         const cfg = new BrowserStackConfig(baseOptions, baseConfig, [
-            {
-                phone: { capabilities: { 'appium:app': 'bs://abc' } },
-                browser: { capabilities: { browserName: 'chrome' } },
-            },
+            { alwaysMatch: { platformName: 'iOS' }, firstMatch: [{ 'appium:app': 'bs://abc' }] },
         ] as any)
         expect(cfg.appAutomate).toBe(true)
     })
 
-    it('marks app_automate for regular multiremote (object) with an app cap', () => {
+    it('marks app_automate for multiremote (object form) with an app cap', () => {
         const cfg = new BrowserStackConfig(baseOptions, baseConfig, {
             phone: { capabilities: { 'appium:bundleId': 'com.example' } },
             browser: { capabilities: { browserName: 'chrome' } },
         } as any)
         expect(cfg.appAutomate).toBe(true)
-    })
-
-    it('marks app_automate when alwaysMatch carries an app cap', () => {
-        const cfg = new BrowserStackConfig(baseOptions, baseConfig, [
-            { alwaysMatch: { 'appium:app': 'bs://abc' }, firstMatch: [] },
-        ] as any)
-        expect(cfg.appAutomate).toBe(true)
-    })
-
-    it('marks app_automate when only a W3C firstMatch entry carries an app cap', () => {
-        const cfg = new BrowserStackConfig(baseOptions, baseConfig, [
-            {
-                alwaysMatch: { platformName: 'iOS' },
-                firstMatch: [{ 'appium:app': 'bs://abc' }],
-            },
-        ] as any)
-        expect(cfg.appAutomate).toBe(true)
-    })
-
-    it('marks app_automate when caps use nested appium:options.app', () => {
-        const cfg = new BrowserStackConfig(baseOptions, baseConfig, [
-            { platformName: 'iOS', 'appium:options': { app: 'bs://abc' } },
-        ] as any)
-        expect(cfg.appAutomate).toBe(true)
-    })
-
-    it('still marks automate when capabilities omit any app signal', () => {
-        const cfg = new BrowserStackConfig(baseOptions, baseConfig, [
-            { browserName: 'chrome' },
-            { browserName: 'firefox' },
-        ] as any)
-        expect(cfg.appAutomate).toBe(false)
-        expect(cfg.automate).toBe(true)
-    })
-
-    it('falls back to options-only behavior when capabilities are omitted', () => {
-        const cfg = new BrowserStackConfig(baseOptions, baseConfig)
-        expect(cfg.appAutomate).toBe(false)
-        expect(cfg.automate).toBe(true)
     })
 })
